🎨 add context and hooks for navigation menu and menu items
