+++ packages/kitchn/src/hooks/useNavigationMenu.ts
@@ -1,0 +1,15 @@
+import React from "react";
+
+import { NavigationMenuContext } from "../contexts/NavigationMenu";
+
+const useNavigationMenu = () => {
+  const context = React.useContext(NavigationMenuContext);
+  if (!context) {
+    throw new Error(
+      "Navigation menu components must be used within NavigationMenu.Container",
+    );
+  }
+  return context;
+};
+
+export default useNavigationMenu;
